deployment/requirements.txt
tensorflow==2.15.1
streamlit
numpy
matplotlib
Pillow
scikit-learn